Fix preparation of auth adapter
